jsdialog: change sendDialogEvent API parameter
WindowId with 64bit length is needed

Reviewed-on: https://gerrit.libreoffice.org/c/online/+/97107
Tested-by: Jenkins CollaboraOffice <jenkinscollaboraoffice@gmail.com>

+++ kit/ChildSession.cpp
@@ -1435,7 +1435,7 @@ bool ChildSession::dialogEvent(const char* /*buffer*/, int /*length*/, const Str
 
     getLOKitDocument()->setView(_viewId);
 
-    unsigned nLOKWindowId = std::stoi(tokens[1].c_str());
+    unsigned long long int nLOKWindowId = std::stoull(tokens[1].c_str());
     getLOKitDocument()->sendDialogEvent(nLOKWindowId,
         tokens.cat(' ', 2).c_str());
